% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRENDA}
\alias{BRENDA}
\title{BRENDA tissue ontology obo file. using version 2021/10}
\format{
A data frame with columns `ID` and `Name`.
}
\usage{
BRENDA
}
\description{
BTO:0000000 is set to be "unknown"
}
\keyword{datasets}
