% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRENDA.R
\name{BRENDAAncestors}
\alias{BRENDAAncestors}
\title{BRENDAAncestors}
\usage{
BRENDAAncestors(x, obo = BRENDAobo)
}
\arguments{
\item{x}{A BRENDA term id, term name, or synonym of a term}

\item{obo}{Optional. The ontology file. BRENDA OBO file will be loaded if not provided.}
}
\value{
If input is a valid BRENDA term, return a dataframe of the term ids
and term names of the input's ancestors.
}
\description{
Get all the ancestors of a term. Parents of its parents will
be retrieved. Based on `get_ancestors()` function in `ontologyIndex`
updated 07/2020
}
\examples{
BRENDAAncestors("BTO:0000034")
BRENDAAncestors("SUM149 cell") # synonym; function will change it to the current name
BRENDAAncestors("BTO:0000216") # input valid but no result, gives character0
}
