% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRENDA.R
\name{BRENDACurrentName}
\alias{BRENDACurrentName}
\title{BRENDACurrentName}
\usage{
BRENDACurrentName(x, obo = BRENDAobo)
}
\arguments{
\item{x}{A BRENDA term id, term name, or synonym of a term}

\item{obo}{Optional. The ontology file. BRENDA OBO file will be loaded if not provided.}
}
\value{
If input is a valid BRENDA term, return a character of the current
term id and term name
}
\description{
Get the current BRENDA term id and name from a term id,
term name, or synonym.
updated 07/2020
}
\examples{
BRENDACurrentName("BTO:0005475") # valid current name, output itself
BRENDACurrentName("SUM149 cell") # valid synonym; change to current name
BRENDACurrentName("BTO:0005388") # valid synonym; change to current name
}
