% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRENDA.R
\name{BRENDADescendants}
\alias{BRENDADescendants}
\title{BRENDADescendants}
\usage{
BRENDADescendants(x, obo = BRENDAobo)
}
\arguments{
\item{x}{A BRENDA term id, term name, or synonym of a term.}

\item{obo}{Optional. The ontology file. BRENDA OBO file will be loaded if not provided.}
}
\value{
If input is a valid BRENDA term, return a dataframe of the term ids.
and term names of the input's descendants.
}
\description{
Get all the descendants of a term. Children of its children will
be retrieved. Based on `get_descendants()` function in `ontologyIndex`
updated 07/2020
}
\examples{
BRENDADescendants("BTO:0000034")
BRENDADescendants("BTO:0000004") # input valid but no result, gives character0
}
