% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRENDA.R
\name{BRENDADevelopFrom}
\alias{BRENDADevelopFrom}
\title{BRENDADevelopFrom}
\usage{
BRENDADevelopFrom(x, obo = BRENDAobo)
}
\arguments{
\item{x}{A BRENDA term id, term name, or synonym of a term.}

\item{obo}{Optional. The ontology file. BRENDA OBO file will be loaded if not provided.}
}
\value{
If input is a valid BRENDA term, return a character of the term id
and term name the input is developed from.
}
\description{
Get the tissue a term is developed from, if available
updated 07/2020
}
\examples{
BRENDADevelopFrom("BTO:0000995")
BRENDADevelopFrom("BTO:0000150")
BRENDADevelopFrom("BTO:0000187") # input is a valid BRENDA term but
# it is not developed from any other tissues, gives character0
}
