% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmS_searchSampleType.R
\name{BRENDAExistName}
\alias{BRENDAExistName}
\title{search for a valid BRENDA ontology term to use for sample type (deprecated 10/2025)}
\usage{
BRENDAExistName(x, file = BRENDA)
}
\arguments{
\item{x}{A name to search for}

\item{file}{The BRENDA data frame, has columns ID and Name}
}
\value{
TRUE or FALSE
}
\description{
Using BRENDA ontology version 2021.
}
\examples{
BRENDAExistName("blood plasma")
BRENDAExistName("random tissue")
}
