% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRENDAobo}
\alias{BRENDAobo}
\title{BRENDA tissue ontology obo file. updated 06/2020}
\format{
An object of class \code{ontology_index} of length 26.
}
\usage{
BRENDAobo
}
\description{
BRENDA tissue ontology obo file. updated 06/2020
}
\keyword{datasets}
