% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRENDA.R
\name{BRENDAPartOf}
\alias{BRENDAPartOf}
\title{BRENDAPartOf}
\usage{
BRENDAPartOf(x, obo = BRENDAobo)
}
\arguments{
\item{x}{A BRENDA term id, term name, or synonym of a term.}

\item{obo}{Optional. The ontology file. BRENDA OBO file will be loaded if not provided.}
}
\value{
If input is a valid BRENDA term, return a dataframe of the term id
and term name the input is part of. Some terms are part of multiple terms.
}
\description{
Get the tissue a term is part of, if available. The function
only retrieves the information directly saved in the obo file, and will not
automatically find "ancestors" in "part of" relationship.
For example, if A is part of B, B is part of C,
the function will NOT automatically return C in the result of A, unless
the relationship between A and C is specified in the obo file.
updated 07/2020
}
\examples{
BRENDAPartOf("BTO:0000187")
BRENDAPartOf("myeloblast")
BRENDAPartOf("BTO:0001732")
}
