% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRENDA.R
\name{BRENDASearch}
\alias{BRENDASearch}
\title{BRENDASearch}
\usage{
BRENDASearch(x, file = BRENDA, contain_all = TRUE)
}
\arguments{
\item{x}{A string or character vector to search for. Multiple search terms
are separated by space.}

\item{file}{The BRENDA data frame, has columns ID and Name.}

\item{contain_all}{if TRUE, will only return the results contain all search terms.
if FALSE, will return results contain any of the given pattern.}
}
\value{
matrix including search result
}
\description{
search for BRENDA tissue name
updated 02/2024
}
\examples{
# search for results that contain all of "HEK" "293" and "T":
BRENDASearch("HEK 293 T", contain_all = TRUE)
BRENDASearch(c("HEK", "293", "T"), contain_all = TRUE)
BRENDASearch(c("HEK 293", "T"), contain_all = TRUE)

# search for results that contain any of "HEK", "SUM" or "HeLa":
BRENDASearch("HEK SUM HeLa", contain_all = FALSE)
BRENDASearch(c("HEK", "SUM", "HeLa"), contain_all = FALSE)
BRENDASearch(c("HEK SUM", "HeLa"), contain_all = FALSE)

}
