% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GEOplatform}
\alias{GEOplatform}
\title{GEO platform names. updated 02/2024}
\format{
An object of class \code{data.frame} with 25828 rows and 4 columns.
}
\usage{
GEOplatform
}
\description{
GPLXXXXX is set to be "unknown"
}
\keyword{datasets}
