% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOplatform.R
\name{GEOPlatformSearch}
\alias{GEOPlatformSearch}
\title{GEOPlatformSearch}
\usage{
GEOPlatformSearch(
  x,
  species = NULL,
  file = GEOplatform,
  contain_all = TRUE,
  accession_only = FALSE
)
}
\arguments{
\item{x}{a string or character vector to search for. input character vector if searching for multiple terms.}

\item{species}{species name to narrow down the search. partial input of an organism name is allowed. e.g. results for species "Homo sapiens" will be pulled if input "homo".}

\item{file}{dataframe for platform infomation. contains "Title" to search for and "Taxnomy" for organism, and "Accession" for platform name. Optional. GEOplatform file will be loaded if not provided.}

\item{contain_all}{if TRUE, results contain all search terms will be returned. if FALSE, results contain any of the given term will be returned.}

\item{accession_only}{if TRUE, only accession id of the search will be returned as a character. if FALSE, the complete information of the platforms will be returned as a dataframe.}
}
\value{
character or dataframe of search result
}
\description{
search for GEO platform name
updated 08/2020
}
\examples{
GEOPlatformSearch("illumina")
GEOPlatformSearch("illumina", species = "homo sapiens")
GEOPlatformSearch(c("Affymetrix", "Transcriptome"), species = "mus", contain_all = TRUE)
}
