% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchAssayType.R
\name{OmicS_searchAssayType}
\alias{OmicS_searchAssayType}
\title{search for an assay type to use}
\usage{
OmicS_searchAssayType(x = "")
}
\arguments{
\item{x}{a string to search for. if empty, will return all available assay types.}
}
\value{
search result
}
\description{
updated 10/2025
}
\examples{
OmicS_searchAssayType()
OmicS_searchAssayType("transcript")
}
