% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmS_searchOrganism.R
\name{OmicS_searchOrganism}
\alias{OmicS_searchOrganism}
\title{search for a pre-defined organism name to use (deprecated 10/2025)}
\usage{
OmicS_searchOrganism(x = "", organism = predefined_organisms)
}
\arguments{
\item{x}{a string to search for (case-insensitive). if empty, will return all available organisms.}

\item{organism}{pre-defined organism character variable.}
}
\value{
character of search result
}
\description{
search for a pre-defined organism name to use (deprecated 10/2025)
}
\examples{
OmicS_searchOrganism()
OmicS_searchOrganism("homo")
}
