% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmS_searchSampleType.R
\name{OmicS_searchSampleType}
\alias{OmicS_searchSampleType}
\title{OmicS_searchSampleType}
\usage{
OmicS_searchSampleType(x, file = BRENDA, contain_all = TRUE)
}
\arguments{
\item{x}{A string or character vector to search for (case-insensitive).
Multiple search terms can be separated by space, or input them as a character vector.}

\item{file}{The BRENDA data frame, has columns ID and Name.}

\item{contain_all}{if TRUE, will only return the results contain all search terms.
if FALSE, will return results contain any of the given pattern.}
}
\value{
matrix including search result
}
\description{
search for BRENDA tissue name
updated 02/2024
}
\examples{
# search for results that contain all of "HEK" "293" and "T":
OmicS_searchSampleType("HEK 293 T", contain_all = TRUE)
OmicS_searchSampleType(c("HEK", "293", "T"), contain_all = TRUE)
OmicS_searchSampleType(c("HEK 293", "T"), contain_all = TRUE)

# search for results that contain any of "HEK", "SUM" or "HeLa":
OmicS_searchSampleType("HEK SUM HeLa", contain_all = FALSE)
OmicS_searchSampleType(c("HEK", "SUM", "HeLa"), contain_all = FALSE)
OmicS_searchSampleType(c("HEK SUM", "HeLa"), contain_all = FALSE)

}
