% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predefined_organisms}
\alias{predefined_organisms}
\title{predefined organism list}
\format{
A character vector.
}
\usage{
predefined_organisms
}
\description{
a list of predefined organism names
}
\keyword{datasets}
