% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predefined_platforms}
\alias{predefined_platforms}
\title{predefined platform list}
\format{
A character vector.
}
\usage{
predefined_platforms
}
\description{
a list of predefined platform names
}
\keyword{datasets}
