% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{readJson}
\alias{readJson}
\title{read an OmicSignature object from json txt file created by writeJson()}
\usage{
readJson(filename)
}
\arguments{
\item{filename}{json file name to read in}
}
\value{
OmicSignature object
}
\description{
To avoid confusion, in the json text file, assume the column 
names in signature dataframe and difexp dataframe have prefix "sig_" 
and "difexp". This corresponds to writeJson() function. 
updated 02/2024
}
