% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceDifexpCol.R
\name{replaceDifexpCol}
\alias{replaceDifexpCol}
\title{Standardize the column names of difexp}
\usage{
replaceDifexpCol(x)
}
\arguments{
\item{x}{A single character vector specifying the column names of the differential
expression dataframe or matrix. Alternatively, the differential
expression dataframe or matrix itself. If the latter, its column name will be
modified and the matrix will be returned.}
}
\value{
standardized column names or standardized matrix
}
\description{
A helper function that rename some of the commonly used columns
of a differential expression dataframe or matrix.
(updated 10/2024)
}
