% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceSigCol.R
\name{replaceSigCol}
\alias{replaceSigCol}
\title{rename the columns of a signature dataframe
updated 02/2024}
\usage{
replaceSigCol(sigdf, long = FALSE)
}
\arguments{
\item{sigdf}{signature dataframe}

\item{long}{if TRUE, replace "symbol" to "signature_symbol", mainly
used to facilitate write into json format; if FALSE, replace "signature_symbol"
to "symbol", i.e. standard sigdf names.}
}
\value{
signature dataframe with column names changed to standard
}
\description{
rename the columns of a signature dataframe
updated 02/2024
}
