% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchAssayType.R
\name{searchAssayType}
\alias{searchAssayType}
\title{search for an assay type to use}
\usage{
searchAssayType(x = "")
}
\arguments{
\item{x}{a string to search for. if empty, will return all available assay types.}
}
\value{
search result
}
\description{
updated 10/2025
}
\examples{
searchAssayType()
searchAssayType("transcript")
}
