% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchOrganism.R
\name{searchOrganism}
\alias{searchOrganism}
\title{searchOrganism}
\usage{
searchOrganism(x, organism = predefined_organisms)
}
\arguments{
\item{x}{a string or character vector to search for. input character vector if searching for multiple terms.}

\item{organism}{pre-defined organism character variable.}
}
\value{
character of search result
}
\description{
search for organism
updated 05/2025
}
\examples{
searchOrganism("Homo")
}
