% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchPlatform.R
\name{searchPlatform}
\alias{searchPlatform}
\title{search for a platform name to use}
\usage{
searchPlatform(x = "", platforms = predefined_platforms, contain_all = TRUE)
}
\arguments{
\item{x}{a string or character vector to search for. input a character vector if searching for multiple terms. if empty, will return all available organisms.}

\item{platforms}{pre-defined platform information character variable.}

\item{contain_all}{if TRUE, results contain all search terms will be returned. if FALSE, results contain any of the given term will be returned.}
}
\value{
character or dataframe of search result
}
\description{
updated 05/2025
}
\examples{
searchPlatform()
searchPlatform("proteomics")
searchPlatform(c("transcript", "single-cell"), contain_all = TRUE)
}
