% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testRepo.R
\name{testRepo}
\alias{testRepo}
\title{a test function}
\usage{
testRepo(num = NULL)
}
\arguments{
\item{num}{A number}
}
\value{
A list
}
\description{
type a number, and it will gives you its square and cubic value.
Or a text, it will gives you the length of it.
}
