% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwriteJson.R
\name{writeJson}
\alias{writeJson}
\title{write OmicSignature object into json txt format}
\usage{
writeJson(OmicObj, file)
}
\arguments{
\item{OmicObj}{A OmicSignature object}

\item{file}{file name to write}
}
\description{
To avoid confusion, in the written json text file, the column
names in signature dataframe and difexp dataframe will have prefix "sig_"
and "difexp" added. This corresponds to readJson() function.
updated 10/2024
}
