% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRENDA}
\alias{BRENDA}
\title{BRENDA tissue ontology obo file. using version 2021/10}
\format{
An object of class \code{data.frame} with 6565 rows and 2 columns.
}
\usage{
BRENDA
}
\description{
BTO:0000000 is set to be "unknown"
}
\keyword{datasets}
