% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSampleType.R
\name{BRENDAExistName}
\alias{BRENDAExistName}
\title{BRENDAExistName}
\usage{
BRENDAExistName(x, file = BRENDA)
}
\arguments{
\item{x}{A name to search for}

\item{file}{The BRENDA data frame, has columns ID and Name}
}
\value{
TRUE or FALSE
}
\description{
search if a string is a valid BRENDA tissue name
updated 02/2024
}
\examples{
BRENDAExistName("blood plasma")
BRENDAExistName("random tissue")
}
