% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRENDA.R
\name{BRENDASearch}
\alias{BRENDASearch}
\title{BRENDASearch}
\usage{
BRENDASearch(x, obo = BRENDAobo, contain_all = TRUE)
}
\arguments{
\item{x}{A string or character vector to search for.
x will be divided into different search patterns by space. See examples below.}

\item{obo}{The ontology file or a character vector of all possible term names. Optional. BRENDA OBO file will be loaded if not provided.}

\item{contain_all}{if TRUE, will only return the results contain all search terms.
if FALSE, will return results contain any of the given pattern.}
}
\value{
matrix including search result
}
\description{
search for ontology term in an obo file or a vector
of all ontology term name
updated 07/2020
}
\examples{
BRENDASearch("MDA MB cell", contain_all = TRUE) 
# search for results that contain "MDA" "MB" and "cell"
BRENDASearch(c("MDA MB", "cell"), contain_all = TRUE) 
# same as above
BRENDASearch(c("MDA-MB", "SUM"), contain_all = FALSE) 
# search for results that contain either "MDA-MB" or "SUM"
}
