% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMetadata.R
\name{createMetadata}
\alias{createMetadata}
\title{template for creating a metadata list for an OmicSignature R6 object
updated 09/2020}
\usage{
createMetadata(
  signature_name,
  organism,
  phenotype = "unknown",
  covariates = "none",
  platform,
  direction_type,
  sample_type = NULL,
  signature_collection = NULL,
  author = NULL,
  year = NULL,
  PMID = NULL,
  keywords = NULL,
  description = NULL,
  category_num = NULL,
  logFC_cutoff = NULL,
  p_value_cutoff = NULL,
  fdr_cutoff = NULL,
  score_cutoff = NULL,
  cutoff_description = NULL,
  other = NULL
)
}
\arguments{
\item{signature_name}{name of the signature.}

\item{organism}{organism. e.g. "Homo Sapiens", "Mus Musculus".}

\item{phenotype}{the phenotype of the signature. e.g. "Gene KO", "Ketodiet", "Parkinson disease", "unknown".}

\item{covariates}{optional but highly recommended. any covariates used in the study. e.g. a signature of aging with covariate "gender".}

\item{platform}{GEO platform. e.g. "GPL11154" is for Illumina HiSeq 2000 Homo sapiens, "GPL570" is for Affymetrix Human Genome U133 Plus 2.0 Array}

\item{direction_type}{the direction information of the signature.
"uni" or "uni-directional" if signature has only one direction or no direction infomation.
"bi" or "bi-directional" if signature contains "up" and "down" regulated features.
"multi" or "multi-directional" if the signature contains more categories.}

\item{sample_type}{optional but highly recommended. a cell line or tissue from BRENDA ontology.}

\item{signature_collection}{optional. collection of which the signature belongs to.}

\item{author}{optional. the author name if the signature is from a published article.}

\item{year}{optional. the year when the signature was created or published.}

\item{PMID}{optional. the PubMed ID if the signature is from a published article.}

\item{keywords}{optional. key words for the signature. examples are "longevity", "perturbation". "drug".}

\item{description}{optional. free text to describe the signature.}

\item{category_num}{optional, specifically used for multi-directional signature to specify how many categories or class the signature contains.}

\item{logFC_cutoff}{optional. log fold change cutoff used to generate the signature, if applicable.}

\item{p_value_cutoff}{optional. p value cutoff used to generate the signature, if applicable.}

\item{fdr_cutoff}{optional. fdr cutoff used to generate the signature, if applicable.}

\item{score_cutoff}{optional. score cutoff used to generate the signature, if applicable.}

\item{cutoff_description}{optional. discription of the cutoff, if applicable.}

\item{other}{optional. a `list` to specify any other user-defined metadata fields.}
}
\value{
a metadata list to create an OmicSignature R6 object
}
\description{
template for creating a metadata list for an OmicSignature R6 object
updated 09/2020
}
