% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GEOplatform}
\alias{GEOplatform}
\title{all current GEO platform names. updated 06/2020}
\format{
An object of class \code{data.frame} with 21227 rows and 9 columns.
}
\usage{
GEOplatform
}
\description{
all current GEO platform names. updated 06/2020
}
\keyword{datasets}
