% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicSignature.R
\name{OmicSignature}
\alias{OmicSignature}
\title{OmicSignature R6 object}
\description{
a R6 object to store signatures generated from experiments. In cluding metadata, signature, and an optional differential expression matrix.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{metadata}}{a list to describe the metadata}

\item{\code{signature}}{a dataframe contains symbol, score (optional) and direction (optional)}

\item{\code{difexp}}{a dataframe for differential expression result}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OmicSignature$new()}}
\item \href{#method-print}{\code{OmicSignature$print()}}
\item \href{#method-extract.signature}{\code{OmicSignature$extract.signature()}}
\item \href{#method-clone}{\code{OmicSignature$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new OmicSignature object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignature$new(metadata, signature, difexp = NULL, print_message = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{required, must be a list}

\item{\code{signature}}{required, must be a dataframe}

\item{\code{difexp}}{optional}

\item{\code{print_message}}{use TRUE if want to see all messages printed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print an OmicSignature object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignature$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract.signature"></a>}}
\if{latex}{\out{\hypertarget{method-extract.signature}{}}}
\subsection{Method \code{extract.signature()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignature$extract.signature(conditions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conditions}}{conditions for new signatures}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a dataframe of new signatures
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmicSignature$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
