% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predefined_assaytypes}
\alias{predefined_assaytypes}
\title{predefined assay type list}
\format{
An object of class \code{character} of length 7.
}
\usage{
predefined_assaytypes
}
\description{
a list of predefined assay type names
}
\keyword{datasets}
