% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{readJson}
\alias{readJson}
\title{read an OmicSignature object from json txt file created by writeJson()
updated 01/2020}
\usage{
readJson(filename)
}
\arguments{
\item{filename}{json file name to read in}
}
\value{
OmicSignature object
}
\description{
read an OmicSignature object from json txt file created by writeJson()
updated 01/2020
}
