% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceDifexpCol.R
\name{replaceDifexpCol}
\alias{replaceDifexpCol}
\title{rename the columns of a differential analysis matrix for OmicSignature R6 object
updated 08/2020}
\usage{
replaceDifexpCol(colname)
}
\arguments{
\item{colname}{a character vector, especially the column name of the differential analysis matrix, i.e. colnames(difexp)
or, if it's the difexp matrix itself, its column name will be modified and the matrix will be returned.}
}
\value{
standardized column names of the difexp matrix before creating the OmicSignature R6 object
}
\description{
rename the columns of a differential analysis matrix for OmicSignature R6 object
updated 08/2020
}
