% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchOrganism.R
\name{searchOrganism}
\alias{searchOrganism}
\title{search for a pre-defined organism name to use}
\usage{
searchOrganism(x = "", organism = predefined_organisms)
}
\arguments{
\item{x}{a string to search for (case-insensitive). if empty, will return all available organisms.}

\item{organism}{pre-defined organism character variable.}
}
\value{
character of search result
}
\description{
updated 05/2025
}
\examples{
searchOrganism()
searchOrganism("homo")
}
