% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOplatform.R
\name{searchPlatform}
\alias{searchPlatform}
\title{searchPlatform}
\usage{
searchPlatform(
  x,
  organism = NULL,
  file = GEOplatform,
  contain_all = TRUE,
  accession_only = FALSE
)
}
\arguments{
\item{x}{a string or character vector to search for. input character vector if searching for multiple terms.}

\item{organism}{organism name to narrow down the search. partial input of an organism name is allowed. e.g. results for organism "Homo sapiens" will be pulled if input "homo".}

\item{file}{GEO platform information dataframe. columns: "Accession", "Name", "Technology", "Organism".}

\item{contain_all}{if TRUE, results contain all search terms will be returned. if FALSE, results contain any of the given term will be returned.}

\item{accession_only}{if TRUE, only accession id of the search will be returned as a character. if FALSE, the complete information of the platforms will be returned as a dataframe.}
}
\value{
character or dataframe of search result
}
\description{
search for GEO platform name
updated 09/2024
}
\examples{
searchPlatform("illumina")
searchPlatform("illumina", organism = "Homo sapiens")
searchPlatform(c("Affymetrix", "Transcriptome"), organism = "mus", contain_all = TRUE)
}
