% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureVecToDF.R
\name{signatureVecToDF}
\alias{signatureVecToDF}
\title{change a signature vector into a dataframe to be saved into OmicSignature object
updated 10/2024}
\usage{
signatureVecToDF(input, bi_directional = FALSE)
}
\arguments{
\item{input}{a character vector of the symbol of the signature, or a numeric vector which names are signature symbols and values are scores.}

\item{bi_directional}{logical.}
}
\value{
signature dataframe with columns "feature_symbol", along with "score" and "direction" if applicable
}
\description{
change a signature vector into a dataframe to be saved into OmicSignature object
updated 10/2024
}
\examples{
signatures <- c("gene1", "gene2", "gene3")
signatureVecToDF(signatures)

signatures <- c(0.45, -3.21, 2.44)
names(signatures) <- c("gene1", "gene2", "gene3")
signatureVecToDF(signatures, bi_directional = TRUE)
}
