% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeSigDF.R
\name{standardizeSigDF}
\alias{standardizeSigDF}
\title{remove missing and duplicated entries in signature dataframe
updated 04/2024}
\usage{
standardizeSigDF(sigdf)
}
\arguments{
\item{sigdf}{signature dataframe}
}
\value{
signature dataframe with empty, duplicate rows removed and ordered by score
}
\description{
remove missing and duplicated entries in signature dataframe
updated 04/2024
}
