% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeSigDF.R
\name{standardizeSigDF}
\alias{standardizeSigDF}
\title{Standardize signature data frame}
\usage{
standardizeSigDF(sigdf)
}
\arguments{
\item{sigdf}{signature dataframe}
}
\value{
signature dataframe with empty, duplicate rows removed and ordered by score
}
\description{
remove missing and duplicated symbols.
updated 08/2024.
}
