% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeSigDF.R
\name{standardizeSigDF}
\alias{standardizeSigDF}
\title{remove empty, duplicate rows in signature dataframe
updated 08/2020}
\usage{
standardizeSigDF(sigdf)
}
\arguments{
\item{sigdf}{signature dataframe}
}
\value{
signature dataframe with empty, duplicate rows removed and ordered by score
}
\description{
remove empty, duplicate rows in signature dataframe
updated 08/2020
}
