% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{writeJson}
\alias{writeJson}
\title{write OmicSignature object into json txt format
updated 01/2020}
\usage{
writeJson(OmicObj, file)
}
\arguments{
\item{OmicObj}{A OmicSignature object}

\item{file}{export file name}
}
\value{
a "finished" message
}
\description{
write OmicSignature object into json txt format
updated 01/2020
}
